+++ ctds/tests/test_connection_timeout.py
@@ -7,11 +7,17 @@ class TestConnectionTimeout(TestExternalDatabase):
     '''Unit tests related to the Connection.timeout attribute.
     '''
 
+    @property
+    def supports_timeout_set(self):
+        return self.freetds_version >= (1, 0, 0)
+
     def test___doc__(self):
         self.assertEqual(
             ctds.Connection.timeout.__doc__,
             '''\
 The connection timeout.
+
+.. note:: Setting the timeout requires FreeTDS version 1.00 or later.
 '''
         )
 
@@ -23,24 +29,21 @@ The connection timeout.
             self.assertEqual(connection.timeout, None)
             connection.timeout = 123
         except ctds.InterfaceError:
-            pass
-        except ctds.NotSupportedError: # pragma: nocover
+            # An interface error is expected if the connection is closed.
             pass
         else:
             self.fail('.timeout did not fail as expected') # pragma: nocover
 
     def test_getset(self):
-        with self.connect(timeout=1) as connection:
-            try:
-                self.assertEqual(connection.timeout, 1)
+        with self.connect(timeout=123) as connection:
+            self.assertEqual(connection.timeout, 123)
+            if self.supports_timeout_set:
                 for timeout in (
-                        int_(123),
-                        long_(123)
+                        int_(321),
+                        long_(321)
                 ):
                     connection.timeout = timeout
                     self.assertEqual(connection.timeout, timeout)
-            except ctds.NotSupportedError: # pragma: nocover
-                pass
 
     def test_timeout(self):
         with self.connect(timeout=1) as connection:
@@ -67,11 +70,9 @@ The connection timeout.
 
         with self.connect(timeout=1) as connection:
             with connection.cursor() as cursor:
-                try:
+                if self.supports_timeout_set:
                     connection.timeout = 2
                     cursor.execute("WAITFOR DELAY '00:00:01.1';SELECT @@VERSION")
-                except ctds.NotSupportedError: # pragma: nocover
-                    pass
 
     def test_typeerror(self):
         with self.connect() as connection: