+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.4] - 2016-05-19
+### Fixed
+- Fix the `ctds.Connection.timeout` property to work with the **DBSETTIME**
+option added in [FreeTDS](http://freetds.org) version *1.00*.
+### Added
+- Support for TDS version *7.4*, introduced with [FreeTDS](http://freetds.org)
+version *1.00*.
+
 ## [1.0.3] - 2016-03-31
 ### Fixed
 - Fix [execute()](https://www.python.org/dev/peps/pep-0249/#execute) and
@@ -34,7 +42,8 @@ _NVARCHAR_ arguments to remote procedure calls.
 ## [1.0.0] - 2016-03-14
 Initial Release
 
-[Unreleased]: https://github.com/zillow/ctds/compare/v1.0.3...HEAD
+[Unreleased]: https://github.com/zillow/ctds/compare/v1.0.4...HEAD
+[1.0.4]: https://github.com/zillow/ctds/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/zillow/ctds/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/zillow/ctds/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/zillow/ctds/compare/v1.0...v1.0.1